requirements.txt
beautifulsoup4==4.12.2
pandas==2.1.0
playwright==1.37.0
python-dotenv==1.0.0
